feat: Add support for GitHub rulesets.
These allow us to block branch creation in the upstream. This could be
accidental where a repo maintainer pushes to `TokTok/$repo` instead of
their fork.

This is in addition to branch protection rules that apply only to
`master`. No branch protection is applied to other branches, but only 2
bots are currently allowed to create branches other than `master`.